+++ Python/yindex.py
@@ -56,6 +56,7 @@ class Yindex:
 	# TODO: remove all files and artifacts on destroy
 	def destroy(self):
 		self.conn.close()
-		os.remove(self.path)
+		if os.path.exists(self.path):
+			os.remove(self.path)
